Remove toArray() and update other methods to no longer use it. Move more tests into separate test files. Update docs to reflect removal of toArray().
