+++ domassist.js
@@ -16,7 +16,6 @@ export { default as matches } from './lib/matches';
 export { default as toggleClass } from './lib/toggleClass';
 export { default as closest } from './lib/closest';
 export { default as ready } from './lib/ready';
-export { default as toArray } from './lib/toArray';
 export { default as styles } from './lib/styles';
 export { default as addAttrs } from './lib/addAttrs';
 export { default as html } from './lib/html';
