+++ lib/hide.js
@@ -1,9 +1,19 @@
-function hide(el) {
-  const currentDisplay = window.getComputedStyle(el).getPropertyValue('display');
+import find from './find';
 
-  if (currentDisplay !== 'none') {
-    el.dataset._currentDisplay = currentDisplay;
-    el.style.display = 'none';
+function hide(selector) {
+  if (Array.isArray(selector)) {
+    selector.forEach((item) => hide(item));
+  }
+  const els = find(selector);
+  if (els.length) {
+    els.forEach((el) => {
+      const currentDisplay = window.getComputedStyle(el).getPropertyValue('display');
+
+      if (currentDisplay !== 'none') {
+        el.dataset._currentDisplay = currentDisplay;
+        el.style.display = 'none';
+      }
+    });
   }
 }
